+++ pisek/utils/pipeline_tools.py
@@ -22,7 +22,7 @@ import sys
 from typing import Callable
 
 from pisek.jobs.job_pipeline import JobPipeline
-from pisek.utils.util import clean_non_relevant_files
+from pisek.utils.util import clean_non_relevant_files, ChangedCWD
 from pisek.utils.text import eprint
 from pisek.utils.terminal import TARGET_LINE_WIDTH
 from pisek.utils.paths import INTERNALS_DIR
@@ -65,18 +65,6 @@ def run_pipeline(path: str, pipeline_class: Callable[[Env], JobPipeline], **env_
 
         clean_non_relevant_files(all_accessed_files)
         return False
-
-
-class ChangedCWD:
-    def __init__(self, path):
-        self._path = path
-
-    def __enter__(self):
-        self._orig_path = os.getcwd()
-        os.chdir(self._path)
-
-    def __exit__(self, exc_type, exc_value, exc_traceback):
-        os.chdir(self._orig_path)
 
 
 class Lock:
