+++ pisek/utils/util.py
@@ -18,16 +18,20 @@ import os
 import shutil
 from typing import Optional
 
-from pisek.jobs.cache import CACHE_CONTENT_FILE
 from pisek.config.task_config import load_config
 from pisek.utils.paths import BUILD_DIR, TESTS_DIR, INTERNALS_DIR
 
 
-def rm_f(fn):
-    try:
-        os.unlink(fn)
-    except FileNotFoundError:
-        pass
+class ChangedCWD:
+    def __init__(self, path):
+        self._path = path
+
+    def __enter__(self):
+        self._orig_path = os.getcwd()
+        os.chdir(self._path)
+
+    def __exit__(self, exc_type, exc_value, exc_traceback):
+        os.chdir(self._orig_path)
 
 
 def _clean_subdirs(task_dir: str, subdirs: list[str]) -> None:
@@ -49,7 +53,6 @@ def is_task_dir(task_dir: str, pisek_directory: Optional[str]) -> bool:
 
 
 def clean_task_dir(task_dir: str, pisek_directory: Optional[str]) -> bool:
-    rm_f(os.path.join(task_dir, CACHE_CONTENT_FILE))
     _clean_subdirs(task_dir, [BUILD_DIR, TESTS_DIR, INTERNALS_DIR])
     return True
